requirements/optional.txt
mistletoe==1.3.0
pandas==2.2.1
pillow==10.2.0
py-abac==0.4.1
traceloop-sdk==0.15.11
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-langchain==0.15.11
opentelemetry-instrumentation-openai==0.15.11
opentelemetry-instrumentation-requests==0.45b0
opentelemetry-instrumentation-transformers==0.15.11
opentelemetry-instrumentation-urllib3==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-semantic-conventions-ai==0.1.1
opentelemetry-util-http==0.45b0
